requirements.txt
pandas
requests
python-telegram-bot==20.7
flask